Moved ngx-vega files + added resize detection
