+++ package.json
@@ -52,6 +52,7 @@
     "blob-stream-browserify": "^0.1.3",
     "cheet.js": "^0.3.3",
     "core-js": "^3.6.5",
+    "css-element-queries": "^1.2.3",
     "d3-scale": "^3.2.1",
     "d3-selection": "^1.4.1",
     "d3-transition": "^1.3.2",
