+++ projects/ngx-vega/src/lib/vega.component.scss
@@ -1,0 +1,7 @@
+:host {
+  div {
+    width: 100%;
+    height: 100%;
+    margin: auto;
+  }
+}
