+++ projects/ngx-vega/src/lib/vega.component.spec.ts
