+++ projects/ngx-vega/src/lib/view-manager.service.ts
